Refactor agentfile_parser.py to fix C0302 linting error
- Extract all dataclasses and utility functions to new agentfile_models.py
- Reduce agentfile_parser.py from 1065 to 704 lines (361 lines saved)
- Fix C0302: Too many lines in module (1065/1000) error
- Achieve perfect 10.00/10 pylint score with no errors or warnings
- All 34 tests pass, functionality fully preserved
- Clean separation of data models from parser logic
